+++ api/commands.py
@@ -1,0 +1,5 @@
+from api.network_map import NetworkMap
+
+commands = dict({
+    'network_map': NetworkMap
+})