+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support for "skip" action on Ikea Symfonisk
 - Support for `rgb` and `rgbw` zigbee groups
 - Decoupled mode support for Xiaomi single key wired switches (QBKG04LM, QBKG11LM)
+- Zigbee Network Map as custom Domoticz page
 
 New devices:
 - Xiaomi MiJia light intensity sensor